requirements.txt
openai==0.27.0
gradio==3.20.0
python-dotenv==1.0.0